+++ app/services/time/agenda.rb
@@ -308,8 +308,8 @@ class Time::Agenda
         weight_row = person_weights[best_pid]
         why_parts = []
         first_name = names_by_person_id[best_pid].to_s.split.first.presence
-        reason_sentence = weight_row&.reason_values&.find(&:positive?)&.sentence
-        why_parts << I18n.t("time.agenda.why.with", name: first_name, reason: reason_sentence.downcase) if first_name && reason_sentence
+        reason_sentence = weight_row&.reason_values&.find(&:positive?)&.clause
+        why_parts << I18n.t("time.agenda.why.with", name: first_name, reason: reason_sentence) if first_name && reason_sentence
         open_detail = nil
         if (standing = standing_rows[best_pid]) && standing[:subject].present?
           days_asked = standing[:wait_days].to_i
@@ -318,7 +318,7 @@ class Time::Agenda
         end
         why_text = why_parts.join(" · ").presence
         next item.with(emphasis: :prep, why: why_text, prep_name: first_name,
-                       prep_detail: open_detail || reason_sentence&.downcase)
+                       prep_detail: open_detail || reason_sentence)
       end
 
       item